Abstracted markers into their own component & filter default to none
