+++ src/components/Map/Map.js
@@ -1,16 +1,14 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 // import { compose, withProps } from 'recompose';
 import {
   GoogleMap,
   LoadScript,
   KmlLayer,
-  Marker,
   // useLoadScript,
 } from '@react-google-maps/api';
-// import { marker } from 'leaflet';
 import UserLocationMarker from '../UserLocationMarker';
+import BusinessMarkers from '../BusinessMarkers';
 import Filter from '../Filter';
-import getSheetData from '../../utils/getSheetData';
 import MainDrawer from '../Drawer/MainDrawer';
 
 const GOOGLE_KEY = process.env.REACT_APP_GOOGLE_API_KEY;
@@ -30,13 +28,7 @@ const mapOptions = {
 };
 
 function Map() {
-  const [markerData, setMarkerData] = useState([]);
   const [category, setCategory] = useState('');
-
-  // Grabbing marker data on render
-  useEffect(() => {
-    getSheetData().then((result) => setMarkerData(result));
-  }, []);
 
   return (
     <LoadScript googleMapsApiKey={GOOGLE_KEY}>
@@ -56,27 +48,7 @@ function Map() {
           options={{ preserveViewport: true }}
         />
         <UserLocationMarker />
-        {/* Markers */}
-        {markerData.map((marker) => {
-          const businessName = marker[0];
-
-          // Coordinates
-          const position = {
-            lat: parseFloat(marker[3]),
-            lng: parseFloat(marker[4]),
-          };
-
-          // Filtering by category
-          // If the filter category is blank, render all
-          return category === '' ? (
-            <Marker key={businessName} position={position} />
-          ) : (
-            // Otherwise conditionally render category
-            marker[1] === category && (
-              <Marker key={businessName} position={position} />
-            )
-          );
-        })}
+        <BusinessMarkers category={category} />
       </GoogleMap>
     </LoadScript>
   );
